% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_alias}
\alias{wd_get_item_alias}
\title{Get the shortest alias of an item.}
\usage{
wd_get_item_alias(item)
}
\arguments{
\item{item}{a wikidata item}
}
\value{
the shortest alias of an item
}
\description{
If no alias found, return the label.
}
\examples{
wd_get_item_alias(item)
}
