% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_instance_of}
\alias{wd_get_item_instance_of}
\title{Get the instance of an item}
\usage{
wd_get_item_instance_of(item)
}
\arguments{
\item{item}{a wikidata item}
}
\value{
the instance of the item
}
\description{
Get the instance of an item
}
\examples{
wd_get_item_instance_of(item)
}
