% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_label}
\alias{wd_get_item_label}
\title{Get the label of a wikidata item
Check for fr first, and get the en otherwise}
\usage{
wd_get_item_label(item)
}
\arguments{
\item{item}{a wikidata item}
}
\value{
the label of the item
}
\description{
Get the label of a wikidata item
Check for fr first, and get the en otherwise
}
\examples{
wd_get_item_label(item)
}
