% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_statement_as_list}
\alias{wd_get_item_statement_as_list}
\title{Get the list of statements for a given property}
\usage{
wd_get_item_statement_as_list(item, prop)
}
\arguments{
\item{item}{a wikidata item}

\item{prop}{the property to get the statement for}
}
\value{
the liste of statements
}
\description{
Get the list of statements for a given property
}
\examples{
wd_get_item_statement_as_list(item,"P527")
}
