% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_statement_as_year}
\alias{wd_get_item_statement_as_year}
\title{Get the year from a statement of the item
(It's very dirty)}
\usage{
wd_get_item_statement_as_year(item, prop)
}
\arguments{
\item{item}{a wikidata item}

\item{prop}{the property to get the year from}
}
\value{
the year inside a statement of the item
}
\description{
Get the year from a statement of the item
(It's very dirty)
}
\examples{
wd_get_item_statement_as_year(item, "P571")
}
