% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_get_item_statement_qualifier_as_list}
\alias{wd_get_item_statement_qualifier_as_list}
\title{Get the qualifiers of statements of the item}
\usage{
wd_get_item_statement_qualifier_as_list(item, prop, qual)
}
\arguments{
\item{item}{a wikidata item}

\item{prop}{the property to get the statements}

\item{qual}{the qualifier to retrieve}
}
\value{
a liste of item statement qualifiers
}
\description{
Get the qualifiers of statements of the item
}
\examples{
wd_get_item_statement_qualifier_as_list(item,"P1365","P585")
}
