% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidata.R
\name{wd_id2url}
\alias{wd_id2url}
\title{Get the wikidata url of an item by id}
\usage{
wd_id2url(id, format = "md")
}
\arguments{
\item{id}{The id of a wikidata item}

\item{format}{The format of the url: "simple" or "md" for markdown ("md" is default)}
}
\value{
The wikidata url of the item
}
\description{
Get the wikidata url of an item by id
}
\examples{
wd_id2url("Q15974764")
}
\references{
\url{https://github.com/cpesr/wikidataESR}
}
\seealso{
\code{\link[WikidataR]{WikidataR}}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
