% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{wdesr.natures}
\alias{wdesr.natures}
\title{Instance_of of french ESR institutions.}
\format{A data frame with 6 variables:
\itemize{
\item id: the wikipedia id of the item;
\item libellé: the libellé (label) of the item;
\item recommandé: whether this item is recommanded to use ("oui") or not ("non");
\item niveau: level of the item (1:international, 2:national, 3:group of institutions, 4:institutions, 5:intermediary level, 6:sub-institutions);
\item wikipedia: url to the wikipedia notice;
\item note: note to help the user.
}}
\source{
\url{https://www.wikidata.org}
}
\usage{
wdesr.natures
}
\description{
A dataset containing the main instance_of (nature) french ESR institutions,
together with some additionnal informations.
}
\keyword{datasets}
