% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{wdesr.status}
\alias{wdesr.status}
\title{Status of french ESR institutions.}
\format{A data frame with 6 variables:
\itemize{
\item id: the wikipedia id of the item;
\item libellé: the libellé (label) of the item;
\item recommandé: whether this item is recommanded to use ("oui") or not ("non");
\item niveau: level of the item (1:national, 2:group of institutions, 3:institutions, 4:intermediary level, 5:sub-institutions, 6:non-institutions);
\item wikipedia: url to the wikipedia notice;
\item note: note to help the user.
}}
\source{
\url{https://www.wikidata.org}
}
\usage{
wdesr.status
}
\description{
A dataset containing the main status (\emph{statuts}) french ESR institutions,
together with some additionnal informations.
}
\keyword{datasets}
