% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{wdesr.statuts}
\alias{wdesr.statuts}
\title{Status of french ESR institutions.}
\format{A data frame with 6 variables:
\itemize{
\item id: the wikipedia id of the item;
\item libellé: the label of the item;
\item recommandé: whether this item is recommanded to use ("oui") or not ("non");
\item niveau: level of the item (see \code{\link{wdesr.niveaux}})
\item wikipedia: url to the wikipedia notice;
\item note: note to help the user.
}}
\source{
\url{https://www.wikidata.org}
}
\usage{
wdesr.statuts

wdesr.statuts
}
\description{
A dataset containing the main status (\emph{statuts}) french ESR institutions,
together with some additionnal informations.

A dataset containing arbitrary levels to institutions statuts.
Used to set sizes when plotting things.
niveau;label;description;exemple
}
\keyword{datasets}
