% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_clear_cache}
\alias{wdesr_clear_cache}
\title{Clear the local WDESR cache}
\usage{
wdesr_clear_cache()
}
\value{
nothing
}
\description{
Must be used whenever the data of wikidata are modified during a R session.
}
\examples{
wdesr_clear_cache()
}
