% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wdesr_get_cache}
\alias{wdesr_get_cache}
\title{Get the cache}
\usage{
wdesr_get_cache()
}
\value{
the environement of the cache
}
\description{
To improve the performances, items are cached.
This function returns this cache.
It contains two dataframes: instances_ofs and items.
}
\examples{
wdesr.cache <- wdesr_get_cache()
wdesr.cache$instance_ofs
wdesr.cache$items
}
\seealso{
\code{\link{wdesr_clear_cache}} \code{\link{wdesr_get_cache}} \code{\link{wdesr_save_cache}} \code{\link{wdesr_load_cache}}
}
