% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_get_data}
\alias{wdesr_get_data}
\title{Loader}
\usage{
wdesr_get_data(wdids)
}
\arguments{
\item{wdids}{a set of wikidata ids}
}
\value{
a dataframe with the data of the universities
}
\description{
Get the data of a set of universities.
Data are cached: use \code{\link{wdesr_clear_cache}} to refresh data from wikidata.
}
\examples{
wdesr_get_data(c("Q3551576","Q2013017"))
}
