% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_get_graph}
\alias{wdesr_get_graph}
\title{Get a graph from a root id, following a given set of properties}
\usage{
wdesr_get_graph(wdid, props, depth = 3, active_only = FALSE)
}
\arguments{
\item{wdid}{the wikidata id of the root}

\item{props}{the set of properties to follow}

\item{depth}{the depth of the graph (more or less) (default to 3)}

\item{active_only}{TRUE to filter dissolved universities (default to FALSE)}
}
\value{
a list of edges and vertices
}
\description{
Get a graph from a root id, following a given set of properties
}
\examples{
wdesr_get_graph("Q61716176",c('composante','associé'), 1)
}
