% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_load_item}
\alias{wdesr_load_item}
\title{Load the data of a university}
\usage{
wdesr_load_item(wdid)
}
\arguments{
\item{wdid}{the wikidata id of the university}
}
\value{
a dataframe with the data of the university
}
\description{
Load the data of a university
}
\examples{
wdesr_load_item("Q3551576")
}
