% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_load_items}
\alias{wdesr_load_items}
\title{Load the data of a set of universities}
\usage{
wdesr_load_items(wdids)
}
\arguments{
\item{wdids}{a set of wikidata ids}
}
\value{
a dataframe with the data of the universities
}
\description{
Load the data of a set of universities
}
\examples{
wdesr_load_items(c("Q3551576","Q2013017"))
}
