% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_log_warnings}
\alias{wdesr_log_warnings}
\title{Output warnings in kable format}
\usage{
wdesr_log_warnings(df.g, kableformat = "pipe")
}
\arguments{
\item{df.g}{A data representing a graph, as returned by wdesr_get_graph.}

\item{kableFormat}{The format used for kableExtra::kable}
}
\value{
A string describing the warnings
}
\description{
Output warnings in kable format
}
\examples{
wdesr_log_warnings(df, "warnings.md")
}
\references{
\itemize{
\item \url{https://github.com/cpesr/wikidataESR}
\item \url{https://www.wikidata.org}
}
}
\seealso{
\code{\link{wdesr_clear_cache}}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
