% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_log_warnings_edges}
\alias{wdesr_log_warnings_edges}
\title{Get warnings for the edges}
\usage{
wdesr_log_warnings_edges(df.g)
}
\arguments{
\item{df.g}{A data representing a graph, as returned by wdesr_get_graph.}
}
\value{
A dataframe with edges warnings
}
\description{
Get warnings for the edges
}
