% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_log_warnings_vertices}
\alias{wdesr_log_warnings_vertices}
\title{Get warnings for the vetices}
\usage{
wdesr_log_warnings_vertices(df.g)
}
\arguments{
\item{df.g}{A data representing a graph, as returned by wdesr_get_graph.}
}
\value{
A dataframe with vetices warnings
}
\description{
Get warnings for the vetices
}
