% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_node_geom}
\alias{wdesr_node_geom}
\title{Get geom_nodeX function.}
\usage{
wdesr_node_geom(node_type = "text")
}
\arguments{
\item{node_type}{Either "text", "text_repel", "label", or "label_repel" (default to "text")}
}
\value{
the suitable geom_nodeX function according the node_type
}
\description{
Get the suitable geom_nodeX function according the node_type
}
\examples{
wdesr_node_geom("text_repel")
}
