% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_node_label_aes}
\alias{wdesr_node_label_aes}
\title{Label builder.}
\usage{
wdesr_node_label_aes(node_label = "alias", alias, label, fondation,
  dissolution)
}
\arguments{
\item{node_label}{either "alias", "alias_date", "long", or "long_date" (default to "alias")}

\item{alias}{the alias of the item}

\item{label}{the label of the item}

\item{fondation}{the foundation date of the item}

\item{dissolution}{the dissolution date of the item}
}
\value{
a label for a node, as a string
}
\description{
Build the label of the nodes.
}
\examples{
node_label_aes("alias", alias, label, fondation, dissolution)
}
