% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wdesr_remove_from_cache}
\alias{wdesr_remove_from_cache}
\title{Remove items from the cache}
\usage{
wdesr_remove_from_cache(...)
}
\arguments{
\item{...}{wikidata ids of the items to remove}
}
\value{
wdesr.cache$items
}
\description{
Remove items from the cache
}
\examples{
wdesr_remove_from_cache("Q30261532","Q30262192")
}
\references{
\itemize{
\item \url{https://github.com/cpesr/wikidataESR}
}
}
\seealso{
\itemize{
\item \code{\link{wdesr_clear_cache}}
\item \code{\link{wdesr_get_cache}}
\item \code{\link{wdesr_save_cache}}
\item \code{\link{wdesr_load_cache}}
\item \code{\link{wdesr_remove_from_cache}}
}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
