% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\docType{package}
\name{wikidataESR}
\alias{wikidataESR}
\alias{wikidataESR-package}
\title{wikidataESR: A package for retrieving and plotting ESR data from wikidata.}
\description{
Wikidata is a convenient tool to model universities information.
This package intends to help using data about the french universities,
namely Enseignement Supérieur et Recherche (ESR).
Its main purpose is to plot graphs about ESR universities relationship.
}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
\item \url{https://www.wikidata.org}
}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
