% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wdesr_clear_cache}
\alias{wdesr_clear_cache}
\title{Clear the local WDESR cache.}
\usage{
wdesr_clear_cache()
}
\value{
The WDESR cache.
}
\description{
To improve the performances, wikidata items are cached.
Must be used whenever the data of wikidata are modified during a R session.
}
\examples{
wdesr_clear_cache()

}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
}
}
\seealso{
\itemize{
\item \code{\link{wdesr_get_cache}}
\item \code{\link{wdesr_save_cache}}
\item \code{\link{wdesr_load_cache}}
}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
