% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_get_data}
\alias{wdesr_get_data}
\title{Loader of universities data.}
\usage{
wdesr_get_data(wdids)
}
\arguments{
\item{wdids}{A set of wikidata ids.}
}
\value{
A dataframe with the data of the universities.
}
\description{
Get the data of a set of universities.
}
\details{
Data are cached: use \code{\link{wdesr_clear_cache}} to refresh data from wikidata.
}
\examples{
items <- wdesr_get_data(c("Q3551576","Q2013017"))

}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
\item \url{https://www.wikidata.org}
}
}
\seealso{
\code{\link{wdesr_clear_cache}}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
