% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_ggplot_graph}
\alias{wdesr_ggplot_graph}
\title{Plot an ESR graph.}
\usage{
wdesr_ggplot_graph(df.g, layout = "kamadakawai", active_only = FALSE,
  node_sizes = c(10, 30), label_sizes = c(4, 6),
  node_label = "alias", node_type = "text", edge_label = TRUE,
  arrow_gap = 0.05, size_guide = FALSE)
}
\arguments{
\item{df.g}{A dataframe representing a graph, as returned by wdesr_get_graph.}

\item{layout}{The layout to use to plot the graph as in \code{\link[sna]{gplot.layout}}.}

\item{active_only}{TRUE to filter the dissolved nodes (default to FALSE).}

\item{node_sizes}{The size of the nodes, either a single value or a range c(min,max).}

\item{label_sizes}{The size of the nodes, either a single value or a range c(min,max).}

\item{node_label}{Define the label for the nodess. Either "alias", "alias_date", "long", or "long_date" (default to "alias").}

\item{node_type}{Define the type of drawing for the nodes. Either "text", "text_repel", "label", or "label_repel" (default to "text").}

\item{edge_label}{TRUE to plot dates on edges (default to "TRUE").}

\item{arrow_gap}{A parameter that will shorten the network edges in order to avoid overplotting edge arrows and nodes see \code{\link[ggnetwork]{fortify.network}}.}

\item{size_guide}{TRUE to plot the guide for sizes (defalut to "FALSE").}
}
\value{
A ggplot2.
}
\description{
A wrapper for ggplot2 to plot graph as returned by \code{\link{wdesr_get_graph}}.
}
\examples{
df.aslace <- wdesr_get_graph("Q61716176",c('composante','associé'), 1)

wdesr_ggplot_graph(df.alsace,
  node_size = c(10,30), label_sizes = c(3,5), arrow_gap = 0.0,
  node_label = "alias", node_type = "text",
  edge_label = FALSE)
}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
\item \url{https://www.wikidata.org}
}
}
\seealso{
\code{\link{wdesr_clear_cache}}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
