% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikidataESR.R
\name{wdesr_load_and_plot}
\alias{wdesr_load_and_plot}
\title{Wrapper to load and plot ESR graphs.}
\usage{
wdesr_load_and_plot(wdid, props = c("composante", "associé"),
  depth = 3, active_only = FALSE, plot_type = "ggplot", ...)
}
\arguments{
\item{wdid}{The wikidata id of the root.}

\item{props}{The properties to follows.}

\item{depth}{The depth of the following}

\item{active_only}{TRUE to filter dissolved universities (default to FALSE).}

\item{plot_type}{Either "ggplot" or "plotly" (default to ggplot).}

\item{...}{Additionnal parameters for the plot; see \code{\link{wdesr_ggplot_graph}} for details.}
}
\value{
A ggplot or a plotly.
}
\description{
Conveniently call \code{\link{wdesr_get_graph}} and then \code{\link{wdesr_ggplot_graph}}.
}
\examples{
wdesr_load_and_plot("Q61716176",c('composante','associé'), 1,
  node_size = c(10,30), label_sizes = c(3,5), arrow_gap = 0.0,
  node_label = "alias", node_type = "text",
  edge_label = FALSE)
}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
\item \url{https://www.wikidata.org}
}
}
\seealso{
\code{\link{wdesr_clear_cache}}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
