% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wdesr_load_cache}
\alias{wdesr_load_cache}
\title{Load the cache}
\usage{
wdesr_load_cache(file = "wdesr-cache.RData", default_cache = FALSE)
}
\value{
the environment of the cache
}
\description{
To improve the performances, items are cached.
This function loads this cache.
}
\examples{
wdesr_load_cache()
}
\seealso{
\code{\link{wdesr_clear_cache}} \code{\link{wdesr_get_cache}} \code{\link{wdesr_save_cache}} \code{\link{wdesr_load_cache}}
}
