% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{wdesr_save_cache}
\alias{wdesr_save_cache}
\title{Save the cache.}
\usage{
wdesr_save_cache(file = "wdesr-cache.RData")
}
\arguments{
\item{file}{The name of the file to save the cache to (default to "wdesr-cache.RData").}
}
\value{
nothing
}
\description{
To improve the performances, wikidata items are cached.
This function saves this cache on the local drive
}
\examples{
wdesr_save_cache()
}
\references{
\itemize{
\item \url{https://github.com/juliengossa/DataESR/tree/master/etablissements.esr/wikidataESR}
}
}
\seealso{
\itemize{
\item \code{\link{wdesr_clear_cache}}
\item \code{\link{wdesr_get_cache}}
\item \code{\link{wdesr_load_cache}}
}
}
\author{
Julien Gossa, \email{gossa@unistra.fr}
}
